+++ Projects/FileEncryption/src/gui/pw_line_edit.cpp
@@ -46,7 +46,7 @@ void PWLineEdit::Clear() {
 
 void PWLineEdit::Extract(Password& pw) {
   QByteArray data = pw_line_->text().toUtf8();
-  int size = data.size();
+  int size = static_cast<int>(data.size());
 
   Lock(data.data(), size);
